+++ examples/next-langchain/app/api/completion-string-output-parser/route.ts
@@ -17,5 +17,5 @@ export async function POST(req: Request) {
 
   const stream = await model.pipe(parser).stream(prompt);
 
-  return new StreamingTextResponse(LangChainAdapter.toAIStream(stream));
+  return new StreamingTextResponse(LangChainAdapter.toDataStream(stream));
 }