+++ examples/next-openai-pages/app/api/call-tool/route.ts
@@ -31,5 +31,5 @@ export async function POST(req: Request) {
     },
   });
 
-  return result.toAIStreamResponse();
+  return result.toDataStreamResponse();
 }