+++ examples/next-openai-pages/pages/api/chat-api-route.ts
@@ -15,5 +15,5 @@ export default async function handler(
 
   // write the AI stream to the response
   // Note: this is sent as a single response, not a stream
-  result.pipeAIStreamToResponse(response);
+  result.pipeDataStreamToResponse(response);
 }