+++ examples/nuxt-openai/server/api/completion.ts
@@ -1,5 +1,5 @@
-import { streamText, StreamingTextResponse, StreamData } from 'ai';
 import { createOpenAI } from '@ai-sdk/openai';
+import { StreamData, streamText } from 'ai';
 
 export default defineLazyEventHandler(async () => {
   const apiKey = useRuntimeConfig().openaiApiKey;
@@ -12,25 +12,22 @@ export default defineLazyEventHandler(async () => {
     // Extract the `prompt` from the body of the request
     const { prompt } = await readBody(event);
 
-    // Ask OpenAI for a streaming chat completion given the prompt
-    const result = await streamText({
-      model: openai('gpt-3.5-turbo'),
-      messages: [{ role: 'user', content: prompt }],
-    });
-
     // optional: use stream data
     const data = new StreamData();
 
     data.append({ test: 'value' });
 
-    // Convert the response into a friendly text-stream
-    const stream = result.toAIStream({
-      onFinal(_) {
+    // Ask OpenAI for a streaming chat completion given the prompt
+    const result = await streamText({
+      model: openai('gpt-3.5-turbo'),
+      messages: [{ role: 'user', content: prompt }],
+      onFinish() {
+        data.append('call completed');
         data.close();
       },
     });
 
     // Respond with the stream
-    return new StreamingTextResponse(stream, {}, data);
+    return result.toDataStreamResponse({ data });
   });
 });