+++ packages/svelte/src/use-chat.ts
@@ -81,8 +81,8 @@ const getStreamedResponse = async (
   previousMessages: Message[],
   abortControllerRef: AbortController | null,
   generateId: IdGenerator,
-  streamMode: 'stream-data' | 'text' | undefined,
-  onFinish: ((message: Message) => void) | undefined,
+  streamProtocol: UseChatOptions['streamProtocol'],
+  onFinish: UseChatOptions['onFinish'],
   onResponse: ((response: Response) => void | Promise<void>) | undefined,
   sendExtraMessageFields: boolean | undefined,
   fetch: FetchFunction | undefined,
@@ -137,7 +137,7 @@ const getStreamedResponse = async (
         tool_choice: chatRequest.tool_choice,
       }),
     },
-    streamMode,
+    streamProtocol,
     credentials: extraMetadata.credentials,
     headers: {
       ...extraMetadata.headers,
@@ -174,6 +174,7 @@ export function useChat({
   experimental_onFunctionCall,
   experimental_onToolCall,
   streamMode,
+  streamProtocol,
   onResponse,
   onFinish,
   onError,
@@ -184,6 +185,11 @@ export function useChat({
   fetch,
   keepLastMessageOnError = false,
 }: UseChatOptions = {}): UseChatHelpers {
+  // streamMode is deprecated, use streamProtocol instead.
+  if (streamMode) {
+    streamProtocol ??= streamMode === 'text' ? 'text' : undefined;
+  }
+
   // Generate a unique id for the chat if not provided.
   const chatId = id || `chat-${uniqueId++}`;
 
@@ -245,7 +251,7 @@ export function useChat({
             get(messages),
             abortController,
             generateId,
-            streamMode,
+            streamProtocol,
             onFinish,
             onResponse,
             sendExtraMessageFields,