+++ packages/solid/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ai-sdk/solid
 
+## 0.0.24
+
+### Patch Changes
+
+- c450fcf7: feat (ui): invoke useChat onFinish with finishReason and tokens
+- e4a1719f: chore (ai/ui): rename streamMode to streamProtocol
+- Updated dependencies [c450fcf7]
+- Updated dependencies [e4a1719f]
+  - @ai-sdk/ui-utils@0.0.21
+
 ## 0.0.23
 
 ### Patch Changes